moved calculate adjoint source funtion to new 'main' script to clean things up combined ASDF and SPECFEM write functions for adjoint source for cleaner implementation re-added util test functions, all passing started on adjsrc test function
